+++ tests/fixtures/external_plugins.py
@@ -1,0 +1,25 @@
+from sys import *
+import sys
+
+
+int = 12
+++int
+
+some_list = list((node for node in 'abc'))
+some_tuple = (1, 2, )
+braking_pycodestyle=3
+
+# commented: str = 'comment'
+
+def camelCase(): ...
+
+'{}'.format(1)
+""
+'%s' % 'test'
+
+assert True
+ipdb.set_trace()
+
+logger.info(
+    'Hello {world}'.format(world='Earth')
+)